refactor(tests): split Python API test suite
